Changes for verrsion 1.2
git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/trunk@20 b456876b-0849-0410-b77d-98878d47e9d5

+++ tiki/tiki-upload_image.php
@@ -27,6 +27,37 @@ $smarty->assign('show','n');
 // Process an upload here
 if(isset($_REQUEST["upload"])) {
   // Check here if it is an upload or an URL
+  $smarty->assign('individual','n');
+  if($userlib->object_has_one_permission($_REQUEST["galleryId"],'image gallery')) {
+    $smarty->assign('individual','y');
+    if($tiki_p_admin != 'y') {
+      // Now get all the permissions that are set for this type of permissions 'image gallery'
+      $perms = $userlib->get_permissions(0,-1,'permName_desc','','image galleries');
+      foreach($perms["data"] as $perm) {
+        $permName=$perm["permName"];
+        if($userlib->object_has_permission($user,$_REQUEST["galleryId"],'image gallery',$permName)) {
+          $$permName = 'y';
+          $smarty->assign("$permName",'y');
+        } else {
+          $$permName = 'n';
+          $smarty->assign("$permName",'n');
+        }
+      }
+    }
+  }
+  if($tiki_p_admin_galleries == 'y') {
+    $tiki_p_view_image_gallery = 'y';
+    $tiki_p_upload_images = 'y';
+    $tiki_p_create_galleries = 'y';
+  }
+
+  if($tiki_p_upload_images != 'y') {
+    $smarty->assign('msg',tra("Permission denied you cannot upload images"));
+    $smarty->display('error.tpl');
+    die;  
+  }
+
+  
   $gal_info = $tikilib->get_gallery($_REQUEST["galleryId"]);
   if($gal_info["thumbSizeX"]==0) $gal_info["thumbSizeX"]=80;
   if($gal_info["thumbSizeY"]==0) $gal_info["thumbSizeY"]=80;  
@@ -64,6 +95,17 @@ if(isset($_REQUEST["upload"])) {
       $error_msg=tra("cannot process upload");
     }
   }
+  $up_thumb = 0;
+  // If the thumbnail was uploaded
+  if(isset($_FILES['userfile2'])&&is_uploaded_file($_FILES['userfile2']['tmp_name'])) {
+      $fp = fopen($_FILES['userfile2']['tmp_name'],"r");
+      $thumb_data = fread($fp,filesize($_FILES['userfile2']['tmp_name']));
+      fclose($fp);
+      $thumb_type = $_FILES['userfile2']['type'];
+      $thumb_size = $_FILES['userfile2']['size'];
+      $thumb_name = $_FILES['userfile2']['name'];
+      $up_thumb = 1;
+  } 
   if(empty($_REQUEST["name"])) {
     $error_msg=tra("You have to provide a name to the image");
   }
@@ -73,30 +115,43 @@ if(isset($_REQUEST["upload"])) {
     die;  
   }
   if(isset($data)) {
-    if(function_exists("ImageCreateFromString")&&(!strstr($type,"gif"))) {
-      $img = imagecreatefromstring($data);
-      $size_x = imagesx($img);
-      $size_y = imagesy($img);
-      // Create thumbnail here 
-      // Use the gallery preferences to get the data
-      $t = imagecreate($gal_info["thumbSizeX"],$gal_info["thumbSizeY"]);
-      $tikilib->ImageCopyResampleBicubic( $t, $img, 0,0,0,0, $gal_info["thumbSizeX"],$gal_info["thumbSizeY"], $size_x, $size_y);
-      // CHECK IF THIS TEMP IS WRITEABLE OR CHANGE THE PATH TO A WRITEABLE DIRECTORY
-      //$tmpfname = 'temp.jpg';
-      $tmpfname = tempnam ("/tmp", "FOO").'.jpg';     
-      imagejpeg($t,$tmpfname);
-      // Now read the information
-      $fp = fopen($tmpfname,"r");
-      $t_data = fread($fp, filesize($tmpfname));
-      fclose($fp);
-      unlink($tmpfname);
-      $t_pinfo = pathinfo($tmpfname);
-      $t_type = $t_pinfo["extension"];
-      $t_type='image/'.$t_type;
-      $imageId = $tikilib->insert_image($_REQUEST["galleryId"],$_REQUEST["name"],$_REQUEST["description"],$name, $type, $data, $size, $size_x, $size_y, $user,$t_data,$t_type);
+    
+    if(!$up_thumb) {
+      if(function_exists("ImageCreateFromString")&&(!strstr($type,"gif"))) {
+        $img = imagecreatefromstring($data);
+        $size_x = imagesx($img);
+        $size_y = imagesy($img);
+        // Create thumbnail here 
+        // Use the gallery preferences to get the data
+        $t = imagecreate($gal_info["thumbSizeX"],$gal_info["thumbSizeY"]);
+        $tikilib->ImageCopyResampleBicubic( $t, $img, 0,0,0,0, $gal_info["thumbSizeX"],$gal_info["thumbSizeY"], $size_x, $size_y);
+        // CHECK IF THIS TEMP IS WRITEABLE OR CHANGE THE PATH TO A WRITEABLE DIRECTORY
+        //$tmpfname = 'temp.jpg';
+        $tmpfname = tempnam ("/tmp", "FOO").'.jpg';     
+        imagejpeg($t,$tmpfname);
+        // Now read the information
+        $fp = fopen($tmpfname,"r");
+        $t_data = fread($fp, filesize($tmpfname));
+        fclose($fp);
+        unlink($tmpfname);
+        $t_pinfo = pathinfo($tmpfname);
+        $t_type = $t_pinfo["extension"];
+        $t_type='image/'.$t_type;
+        $imageId = $tikilib->insert_image($_REQUEST["galleryId"],$_REQUEST["name"],$_REQUEST["description"],$name, $type, $data, $size, $size_x, $size_y, $user,$t_data,$t_type);
+      } else {
+        $tmpfname='';
+        $imageId = $tikilib->insert_image($_REQUEST["galleryId"],$_REQUEST["name"],$_REQUEST["description"],$name, $type, $data, $size, 0, 0, $user,'','');
+      }
     } else {
-      $tmpfname='';
-      $imageId = $tikilib->insert_image($_REQUEST["galleryId"],$_REQUEST["name"],$_REQUEST["description"],$name, $type, $data, $size, 0, 0, $user,'','');
+      if(function_exists("ImageCreateFromString")&&(!strstr($type,"gif"))) {
+        $img = imagecreatefromstring($data);
+        $size_x = imagesx($img);
+        $size_y = imagesy($img);
+      } else {
+        $size_x = 0;
+        $size_y = 0;
+      }
+      $imageId = $tikilib->insert_image($_REQUEST["galleryId"],$_REQUEST["name"],$_REQUEST["description"],$name, $type, $data, $size, $size_x, $size_y, $user,$thumb_data,$thumb_type);
     }
     $smarty->assign_by_ref('imageId',$imageId);
     // Now that the image was inserted we can display the image here.
@@ -112,7 +167,40 @@ if(isset($_REQUEST["galleryId"])) {
 } else {
   $smarty->assign('galleryId','');
 }
-$galleries = $tikilib->list_galleries(0,-1,'lastModif_desc', $user,'');
+if($tiki_p_admin_galleries != 'y') {
+  $galleries = $tikilib->list_visible_galleries(0,-1,'lastModif_desc', $user,'');
+} else {
+  $galleries = $tikilib->list_galleries(0,-1,'lastModif_desc', $user,'');
+}
+for($i=0;$i<count($galleries["data"]);$i++) {
+  if($userlib->object_has_one_permission($galleries["data"][$i]["galleryId"],'image gallery')) {
+    $galleries["data"][$i]["individual"]='y';
+    
+    if($userlib->object_has_permission($user,$galleries["data"][$i]["galleryId"],'image gallery','tiki_p_view_image_gallery')) {
+      $galleries["data"][$i]["individual_tiki_p_view_image_gallery"]='y';
+    } else {
+      $galleries["data"][$i]["individual_tiki_p_view_image_gallery"]='n';
+    }
+    if($userlib->object_has_permission($user,$galleries["data"][$i]["galleryId"],'image gallery','tiki_p_upload_images')) {
+      $galleries["data"][$i]["individual_tiki_p_upload_images"]='y';
+    } else {
+      $galleries["data"][$i]["individual_tiki_p_upload_images"]='n';
+    }
+    if($userlib->object_has_permission($user,$galleries["data"][$i]["galleryId"],'image gallery','tiki_p_create_galleries')) {
+      $galleries["data"][$i]["individual_tiki_p_create_galleries"]='y';
+    } else {
+      $galleries["data"][$i]["individual_tiki_p_create_galleries"]='n';
+    }
+    if($tiki_p_admin=='y' || $userlib->object_has_permission($user,$galleries["data"][$i]["galleryId"],'image gallery','tiki_p_admin_galleries')) {
+      $galleries["data"][$i]["individual_tiki_p_create_galleries"]='y';
+      $galleries["data"][$i]["individual_tiki_p_upload_images"]='y';
+      $galleries["data"][$i]["individual_tiki_p_view_image_gallery"]='y';
+    } 
+    
+  } else {
+    $galleries["data"][$i]["individual"]='n';
+  }
+}
 $smarty->assign_by_ref('galleries',$galleries["data"]);
 
 // Display the template
